refactor(nlp): text-classification tests variables renamed
